Move HM code from XPHM to HM script
